Rename post processor files
